Přidány JSON-JAVA classy - funkční x ještě se nepřevádějí datumy na Date, je to Longint

+++ app/src/main/java/com/example/jirka/retrofit1/JSON/Main.java
@@ -14,7 +14,7 @@ public class Main {
     private Double temp;
     @SerializedName("pressure")
     @Expose
-    private Integer pressure;
+    private Double pressure;
     @SerializedName("humidity")
     @Expose
     private Integer humidity;
@@ -53,7 +53,7 @@ public class Main {
      * @return
      *     The pressure
      */
-    public Integer getPressure() {
+    public Double getPressure() {
         return pressure;
     }
 
@@ -62,11 +62,11 @@ public class Main {
      * @param pressure
      *     The pressure
      */
-    public void setPressure(Integer pressure) {
+    public void setPressure(Double pressure) {
         this.pressure = pressure;
     }
 
-    public Main withPressure(Integer pressure) {
+    public Main withPressure(Double pressure) {
         this.pressure = pressure;
         return this;
     }
